task1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include <stdlib>
+#include <stdlib.h>
 #include "function_pointers.h"
 /**
  * array_iterator - executes a function given as a parameter
@@ -8,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < size; i++)
 		action(array[i]);
